compound Text.Mono
