+++ src/components/texts/Text/Text.stories.tsx
@@ -1,24 +1,22 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import TextComponent from './';
+import { layoutStylePropsController } from '.storybook/stylePropsController';
 
 const loremParagrpah =
   'Ante succurro audacia suus aufero sui aestivus tempus. Subiungo totam adicio. Quasi usque eius clam caries cavus subiungo. Cui corona venustas virgo bos vomer cunabula deficio. Ultra coepi bardus aperte stillicidium. Condico aetas tempore voco iusto deputo. Validus viriliter volva ventus tenax cupressus vix strenuus cupressus. Acervus venia cornu celer. Alveus theologus vulpes video vesper agnitio apostolus curiositas. Valetudo debilito adflicto dolorem suus utrum theatrum. Tolero adopto compello. Spiritus tenuis coaegresco alter artificiose verbum a conqueror tergum aufero. Ventosus animadverto vito absorbeo decumbo censura autem aufero. Dens perferendis laudantium. Asporto cuius carmen peccatus a sto cometes deduco.';
 
-const meta: Meta<typeof TextComponent> = {
-  component: TextComponent,
+const meta: Meta<typeof TextComponent.Default> = {
+  component: TextComponent.Default,
 };
 
 export default meta;
-type Story = StoryObj<typeof TextComponent>;
+type Story = StoryObj<typeof TextComponent.Default>;
 
 export const Text: Story = {
   args: {
     children: loremParagrpah,
   },
   argTypes: {
-    span: {
-      control: { type: 'boolean' },
-    },
     lineClamp: {
       control: { type: 'number' },
     },
@@ -33,5 +31,6 @@ export const Text: Story = {
       options: ['p', 'span', 'a'],
       control: { type: 'select' },
     },
+    ...layoutStylePropsController,
   },
 };
