+++ src/components/texts/Text/monoStyles.ts
